Fix billing_code_type normalization in dim_procedure, resolve accepted_values test failures, and align BigQuery types (varchar→string). All dbt tests passing.

+++ dbt/tests/assert_dim_procedure_pct_unknown_not_high.sql
@@ -8,4 +8,4 @@ with cts as (
 )
 select 1 as warn_high_unknown_pct
 from cts
-where total > 0 and (cast(unknown_count as float) / total) >= 0.95
+where total > 0 and (cast(unknown_count as FLOAT64) / total) >= 0.95
